51 uptodate

+++ routes/web.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Route;
 use App\Models\Listing;
 use App\Http\Controllers\ListingController;
+use App\Http\Controllers\UserController;
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -20,24 +21,38 @@ Route::get('/',[ListingController::class,'index']);
 
 //show create form
 
-Route::get('/listings/create',[ListingController::class,'create']);
+Route::get('/listings/create',[ListingController::class,'create'])
+->middleware('auth');
 
 
 //Store listing data
-Route::post('/listings',[ListingController::class,'store']);
+Route::post('/listings',[ListingController::class,'store'])->middleware('auth');
 
 //Show Edit form
 
-Route::get('/listings/{listing}/edit',[ListingController::class,'edit']);
+Route::get('/listings/{listing}/edit',[ListingController::class,'edit'])->middleware('auth');
 
 //Update edit form
-Route::put('/listings/{listing}',[ListingController::class,'update']);
+Route::put('/listings/{listing}',[ListingController::class,'update'])->middleware('auth');
 
 //Delete listing 
-Route::delete('/listings/{listing}',[ListingController::class,'destroy']);
+Route::delete('/listings/{listing}',[ListingController::class,'destroy'])->middleware('auth');
 //Single Listing
 
 Route::get('/listings/{listing}',
 [ListingController::class,'show']);
 
+// SHow register/Create Form
+Route::get('/register',[UserController::class,'create'])->middleware('guest');
 
+//create new user
+Route::post('/users',[UserController::class,'store']);
+
+//Logout
+Route::post('/logout',[UserController::class,'logout'])->middleware('auth');
+
+//login form
+Route::get('/login',[UserController::class,'login'])->name('login')->middleware('guest');
+
+//login user
+Route::post('/users/authenticate',[UserController::class,'authenticate']);
